# Conflicts:
#	backend/src/main/java/com/example/faishion/image/Image.java
#	backend/src/main/java/com/example/faishion/review/Review.java
#	backend/src/main/java/com/example/faishion/seller/SellerRepository.java
#	backend/src/main/java/com/example/faishion/stock/StockRepository.java

+++ backend/src/main/java/com/example/faishion/stock/StockRepository.java
@@ -2,5 +2,6 @@ package com.example.faishion.stock;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StockRepository extends JpaRepository<Stock, Long> {
+public interface StockRepository extends JpaRepository<Stock, Integer> {
+
 }
